Move team components to sub directory
