+++ frontend/src/components/team/bio.vue
