+++ frontend/src/views/character.vue
@@ -168,7 +168,7 @@ import BISTable from '@/components/bis_table.vue'
 import CharacterBio from '@/components/character_bio.vue'
 import DeleteBIS from '@/components/modals/confirmations/delete_bis.vue'
 import DeleteCharacter from '@/components/modals/confirmations/delete_character.vue'
-import TeamBio from '@/components/team_bio.vue'
+import TeamBio from '@/components/team/bio.vue'
 import BISList from '@/interfaces/bis_list'
 import { CharacterDetails } from '@/interfaces/character'
 import Job from '@/interfaces/job'
