+++ frontend/src/views/team/join.vue
@@ -69,8 +69,8 @@
 import { Component, Vue } from 'vue-property-decorator'
 import CharacterBio from '@/components/character_bio.vue'
 import ClaimCharacterModal from '@/components/modals/confirmations/claim_character.vue'
-import TeamBio from '@/components/team_bio.vue'
-import TeamMemberForm from '@/components/team_member_form.vue'
+import TeamBio from '@/components/team/bio.vue'
+import TeamMemberForm from '@/components/team/member_form.vue'
 import { Character } from '@/interfaces/character'
 import { TeamCreateResponse, TeamMemberUpdateErrors } from '@/interfaces/responses'
 import Team from '@/interfaces/team'
